Arrange Layout for many components, Check Home.jsx
