--- src/components/Test.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function Test() {
-  return <div>Test</div>;
-}
-
-export default Test;
